+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## v2.9.0
-* Add field for custom limit amount and custom error message
+* Add parameter for custom limit amount in forms
 
 ## v2.8.0
 * (re)Fix max total amount to 5000